chatbot_backend/requirements.txt
bcrypt==4.1.2
email-validator==2.1.0.post1
fastapi==0.104.1
Jinja2==3.1.3
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
langchain-google-genai==0.0.9
langsmith==0.1.9
openai==1.33.0
passlib==1.7.2
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
SQLAlchemy==2.0.23
uvicorn==0.24.0

